+++ Entities/DataTransferObject/RateSellerReturn.cs
@@ -9,7 +9,7 @@ namespace Entities.DataTransferObject
         public int RateOwnerId { get; set; }
         public string RateOwnerName { get; set; }
         public string RateOwnerAvatar { get; set; }
-        public int NumOfStar { get; set; }
+        public double NumOfStar { get; set; }
         public string Content { get; set; }
         public string TradingPostTitle { get; set; }
         public int TradingPostId { get; set; }
